.old/templates/cbp-gateway/pyproject.toml
[tool.poetry]
name = "cbp-gateway"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "cbp_gateway"}]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.22"
pyzmq = "^25.1.1"
fastapi = "^0.104.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.209.7"}
pydantic = "^2.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
